--- src/components/Home/EarthquakeTrendChart/Chart.jsx
@@ -1,179 +1,0 @@
-import * as echarts from 'echarts';
-import React, { useEffect, useRef } from 'react';
-
-import { CHART_COLORS } from '@/app/constants/chart';
-
-import { SCALE_CHART_OPTIONS } from './constants';
-
-const Chart = () => {
-  const chartRef = useRef(null);
-
-  const SCALE_CHART_OPTIONS = {
-    color: [
-      CHART_COLORS.chart1,
-      CHART_COLORS.chart2,
-      CHART_COLORS.chart3,
-      CHART_COLORS.chart4
-    ],
-    tooltip: {
-      trigger: 'axis',
-      axisPointer: {
-        type: 'cross',
-        label: {
-          backgroundColor: '#6a7985'
-        }
-      }
-    },
-    legend: {
-      data: ['1년', '3년', '5년', '10년']
-    },
-    toolbox: {
-      feature: {
-        saveAsImage: {}
-      }
-    },
-    grid: {
-      left: '3%',
-      right: '4%',
-      bottom: '3%',
-      containLabel: true
-    },
-    xAxis: [
-      {
-        type: 'category',
-        boundaryGap: false,
-        data: ['규모 3.0 ~', '규모 3.0 ~ 4.0', '규모 4.0 ~ 5.0', '규모 5.0 ~']
-      }
-    ],
-    yAxis: [
-      {
-        type: 'value'
-      }
-    ],
-    series: [
-      {
-        name: '1년',
-        type: 'line',
-        stack: 'Total',
-        smooth: true,
-        lineStyle: {
-          width: 0
-        },
-        showSymbol: false,
-        areaStyle: {
-          opacity: 0.8,
-          color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-            {
-              offset: 0,
-              color: 'rgb(128, 255, 165)'
-            },
-            {
-              offset: 1,
-              color: 'rgb(1, 191, 236)'
-            }
-          ])
-        },
-        emphasis: {
-          focus: 'series'
-        },
-        data: [140, 232, 101, 264, 90, 340, 250]
-      },
-      {
-        name: '3년',
-        type: 'line',
-        stack: 'Total',
-        smooth: true,
-        lineStyle: {
-          width: 0
-        },
-        showSymbol: false,
-        areaStyle: {
-          opacity: 0.8,
-          color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-            {
-              offset: 0,
-              color: 'rgb(0, 221, 255)'
-            },
-            {
-              offset: 1,
-              color: 'rgb(77, 119, 255)'
-            }
-          ])
-        },
-        emphasis: {
-          focus: 'series'
-        },
-        data: [120, 282, 111, 234, 220, 340, 310]
-      },
-      {
-        name: '5년',
-        type: 'line',
-        stack: 'Total',
-        smooth: true,
-        lineStyle: {
-          width: 0
-        },
-        showSymbol: false,
-        areaStyle: {
-          opacity: 0.8,
-          color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-            {
-              offset: 0,
-              color: 'rgb(55, 162, 255)'
-            },
-            {
-              offset: 1,
-              color: 'rgb(116, 21, 219)'
-            }
-          ])
-        },
-        emphasis: {
-          focus: 'series'
-        },
-        data: [320, 132, 201, 334, 190, 130, 220]
-      },
-      {
-        name: '10년',
-        type: 'line',
-        stack: 'Total',
-        smooth: true,
-        lineStyle: {
-          width: 0
-        },
-        showSymbol: false,
-        areaStyle: {
-          opacity: 0.8,
-          color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-            {
-              offset: 0,
-              color: 'rgb(255, 0, 135)'
-            },
-            {
-              offset: 1,
-              color: 'rgb(135, 0, 157)'
-            }
-          ])
-        },
-        emphasis: {
-          focus: 'series'
-        },
-        data: [220, 402, 231, 134, 190, 230, 120]
-      }
-    ]
-  };
-  useEffect(() => {
-    if (chartRef.current) {
-      const chart = echarts.init(chartRef.current);
-
-      chart.setOption(SCALE_CHART_OPTIONS);
-    }
-  }, [SCALE_CHART_OPTIONS, chartRef]);
-  return (
-    <>
-      <div className='h-11' />
-      <div className='h-2/3 w-full' ref={chartRef} />
-    </>
-  );
-};
-
-export default Chart;
