+++ src/components/Home/EarthquakeTrendChart/constants/index.ts
@@ -1,7 +1,6 @@
 export const SELECT_FILTER = [
-    { value: '1', label: '1년' },
-    { value: '3', label: '3년' },
-    { value: '5', label: '5년' },
-    { value: '10', label: '10년' }
-  ];
-
+  { value: '1', label: '1년' },
+  { value: '3', label: '3년' },
+  { value: '5', label: '5년' },
+  { value: '10', label: '10년' }
+];
